% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emCalcMix.R
\name{dfSort}
\alias{dfSort}
\title{dfSort: Utility function to sort a data frame.}
\usage{dfSort(df, trt_var)}
\arguments{
\item{df}{A data frame. Defaults to NULL.}

\item{trt_var}{A variable name, in quotes. Defaults to NULL.}
}
\value{
A list containing the following items:

\item{X}{The data frame sorted in descending order by trt_var.}

}
\description{
This function returns the input data frame, sorted in descending order by trt_var. Used to prep data frame before running funbartCausal() or funbartCausalProbit() functions, which require all inputs to be sorted with treatment=1 first, then treatment=0.
}
\examples{
\dontrun{
## Sort data frame
sorted <- dfSort(mydf, trt_var = "trt")
}}
\author{
Jennifer E. Starling
}
\keyword{funbart}


