% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KStest.R
\name{sim}
\alias{sim}
\title{Simulated cosine data with four covariates.}
\format{
\describe{
\item{y}{The functional response, observed with noise.}
\item{ti}{The time point for each observation.}
\item{fx}{The functional response without noise.}
\item{eps}{Residual error.}
\item{x1, x2, x3, x4}{The four covariates, simulated as pairs of bivariate Gaussians.}
\item{id}{A unique identifier for each observation.}
\item{train}{Train/test set indicators for a random 70-30 split.}
}}
\description{
A simulated data set to illustrate the Functional BART
method.  Data contains four 'x' covariates; each pair
is simulated from bivariate Gaussians with moderate correlation
and unit variance.  Each observation has one corresponding time
point at which it is observed, randomly chosen from eight discrete
time points.

Function values are calculated as
f(x,t) = g(x1,x2) * cos(t + 2 * pi * h(x2,x3)) where g() and h()
are sums of the covariates.

This configuration simulates both amplitude change and phase shift for the
underlying function.
}
\examples{
\dontrun{
# Load data set.
data(sim)
head(sim)
}}
\author{
Jennifer E. Starling
}
\keyword{funbart}
