% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KStest.R
\name{tuneEcross}
\alias{tuneEcross}
\title{tuneEcross: Utility function for tuning expected number of crossings for tsbart().}
\usage{
tuneEcross(ecross_candidates = seq(.25,5,by=.25),
                      y, tgt, tpred, x, xpred, nburn=1000, nsim=1000, ntree=200,
                      lambda=NULL, sigq=.9, sighat=NULL, nu=3,
                      ecross=1, base_tree=.95, power_tree=2,
                      probit=FALSE, yobs=NULL)
}
\arguments{

\item{ecross_candidates}{A vector of candidate values for expected number of crossings. Required, defaults to seq(.25,5,by=.25).}

\item{y}{Length n vector with continuous response values.  In probit case, should contain initializations for latent variables.}

\item{tgt}{Length n targeted covariate over which to smooth.}

\item{tpred}{Length npred out-of-sample targeted covariate.}

\item{x}{A (n x p) data frame or matrix containing the covariates which are not to be smoothed over.}

\item{xpred}{A (npred x p) data frame or matrix containing out of sample covariates.}

\item{nburn}{Number of burn-in MCMC iterations.}

\item{nsim}{Number of MCMC iterations to save after burn-in.}

\item{ntree}{Number of trees. Defaults to 200.}

\item{lambda}{Scale parameter in the chisq prior sigma^2.  Defaults to NULL, ie being estimated from sigq and sighat. Not appicable for binary case.}

\item{sigq}{Calibration quantile in the chisq prior on sigma^2.  Defaults to .9. Not applicable for binary case.}

\item{sighat}{Calibration estimate for chisq prior on sigma^2.  Defaults to being estimated from data using linear model.  Not applicable for binary case.}

\item{nu}{Degrees of freedom in the chisq prior on sigma^2. Defaults to 3. Not applicable for binary case.}

\item{base_tree}{Base for tree prior.  Defaults to 0.95.}

\item{power_tree}{Power for tree prior.  Defaults to 2.0.}

\item{probit}{0 indicates continuous response; 1 indicates probit response.  Default is 0.  If 1, then yobs must be populated.}

\item{yobs}{Length n vector of binary responses; only populated for probit=1 case.}
}
\value{
A list object, containing the following items:

\item{ecross_opt}{A double; the optimal expected crossings value from the user input grid.}

\item{waic_plot}{A ggplot object containing diagnostics for tuning expected crossings.}

\item{waic_grid}{A data frame containing the expected crossing candidates and corresponding WAIC values.}

}
\description{
This function provides tunes the expected number of crossings.  See vignette for detail.
}
\examples{
\dontrun{
## Evaluate optimal number of crossings.
ecrossTune = tuneEcross(ecross_candidates = seq(.25,5,by=.25),
                      y, tgt, tpred, x, xpred, nburn, nsim, ntree=200,
                      lambda=NULL, sigq=.9, sighat=NULL, nu=3,
                      ecross=1, base_tree=.95, power_tree=2,
                      probit=FALSE, yobs=NULL)

## View expected number of crossings.
ecrossTune[["ecross_opt"]]

}}
\author{
Jennifer E. Starling
}
\keyword{tsbart}


